# backend/internal/api/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn4AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization",
	}
	rn12AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn2AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"GET":    "Authorization",
	}
	rn3AllowedHeaders = map[string]string{
		"POST": "Authorization",
	}
	rn17AllowedHeaders = map[string]string{
		"POST": "Authorization",
	}
	rn6AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn9AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"PUT":    "Authorization,Content-Type",
	}
	rn7AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn15AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn11AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"PUT":    "Authorization,Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/"

			if l := len("/api/"); len(elem) >= l && elem[0:l] == "/api/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'b': // Prefix: "blocks"

				if l := len("blocks"); len(elem) >= l && elem[0:l] == "blocks" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleListBlocksRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleCreateBlockRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET,POST",
							allowedHeaders: rn4AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'c': // Prefix: "current"
						origElem := elem
						if l := len("current"); len(elem) >= l && elem[0:l] == "current" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleGetCurrentBlockRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: rn12AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

						elem = origElem
					}
					// Param: "id"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch r.Method {
						case "DELETE":
							s.handleDeleteBlockRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleGetBlockRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "DELETE,GET",
								allowedHeaders: rn2AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'a': // Prefix: "archive"

							if l := len("archive"); len(elem) >= l && elem[0:l] == "archive" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "POST":
									s.handleArchiveBlockByIdRequest([1]string{
										args[0],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "POST",
										allowedHeaders: rn3AllowedHeaders,
										acceptPost:     "",
										acceptPatch:    "",
									})
								}

								return
							}

						case 'u': // Prefix: "unarchive"

							if l := len("unarchive"); len(elem) >= l && elem[0:l] == "unarchive" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "POST":
									s.handleUnarchiveBlockRequest([1]string{
										args[0],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "POST",
										allowedHeaders: rn17AllowedHeaders,
										acceptPost:     "",
										acceptPatch:    "",
									})
								}

								return
							}

						}

					}

				}

			case 'd': // Prefix: "days"

				if l := len("days"); len(elem) >= l && elem[0:l] == "days" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "POST":
						s.handleCreateDayRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn6AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "DELETE":
							s.handleDeleteDayRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PUT":
							s.handleUpdateDayRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "DELETE,PUT",
								allowedHeaders: rn9AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'o': // Prefix: "oidc-config"

				if l := len("oidc-config"); len(elem) >= l && elem[0:l] == "oidc-config" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetOIDCConfigRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 's': // Prefix: "shopping"

				if l := len("shopping"); len(elem) >= l && elem[0:l] == "shopping" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "POST":
						s.handleCreateShoppingItemRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn7AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'r': // Prefix: "reorder"
						origElem := elem
						if l := len("reorder"); len(elem) >= l && elem[0:l] == "reorder" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleReorderShoppingItemsRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "POST",
									allowedHeaders: rn15AllowedHeaders,
									acceptPost:     "application/json",
									acceptPatch:    "",
								})
							}

							return
						}

						elem = origElem
					}
					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "DELETE":
							s.handleDeleteShoppingItemRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PUT":
							s.handleUpdateShoppingItemRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "DELETE,PUT",
								allowedHeaders: rn11AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'u': // Prefix: "userinfo"

				if l := len("userinfo"); len(elem) >= l && elem[0:l] == "userinfo" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetUserInfoRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/"

			if l := len("/api/"); len(elem) >= l && elem[0:l] == "/api/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'b': // Prefix: "blocks"

				if l := len("blocks"); len(elem) >= l && elem[0:l] == "blocks" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = ListBlocksOperation
						r.summary = "ブロック一覧を取得する"
						r.operationID = "listBlocks"
						r.operationGroup = ""
						r.pathPattern = "/api/blocks"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = CreateBlockOperation
						r.summary = "新しいブロックを作成する（アーカイブなし）"
						r.operationID = "createBlock"
						r.operationGroup = ""
						r.pathPattern = "/api/blocks"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'c': // Prefix: "current"
						origElem := elem
						if l := len("current"); len(elem) >= l && elem[0:l] == "current" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = GetCurrentBlockOperation
								r.summary = "現在のブロックを取得する（存在しない場合は自動作成）"
								r.operationID = "getCurrentBlock"
								r.operationGroup = ""
								r.pathPattern = "/api/blocks/current"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

						elem = origElem
					}
					// Param: "id"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch method {
						case "DELETE":
							r.name = DeleteBlockOperation
							r.summary = "ブロックを削除する（最後に作成された未アーカイブのブロックのみ可）"
							r.operationID = "deleteBlock"
							r.operationGroup = ""
							r.pathPattern = "/api/blocks/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							r.name = GetBlockOperation
							r.summary = "ブロックを取得する"
							r.operationID = "getBlock"
							r.operationGroup = ""
							r.pathPattern = "/api/blocks/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'a': // Prefix: "archive"

							if l := len("archive"); len(elem) >= l && elem[0:l] == "archive" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "POST":
									r.name = ArchiveBlockByIdOperation
									r.summary = "指定ブロックをアーカイブする（未アーカイブの中で最古のブロックのみ可）"
									r.operationID = "archiveBlockById"
									r.operationGroup = ""
									r.pathPattern = "/api/blocks/{id}/archive"
									r.args = args
									r.count = 1
									return r, true
								default:
									return
								}
							}

						case 'u': // Prefix: "unarchive"

							if l := len("unarchive"); len(elem) >= l && elem[0:l] == "unarchive" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "POST":
									r.name = UnarchiveBlockOperation
									r.summary = "指定ブロックのアーカイブを解除する（アーカイブ済みの中で最新のブロックのみ可）"
									r.operationID = "unarchiveBlock"
									r.operationGroup = ""
									r.pathPattern = "/api/blocks/{id}/unarchive"
									r.args = args
									r.count = 1
									return r, true
								default:
									return
								}
							}

						}

					}

				}

			case 'd': // Prefix: "days"

				if l := len("days"); len(elem) >= l && elem[0:l] == "days" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "POST":
						r.name = CreateDayOperation
						r.summary = "日エントリを追加する"
						r.operationID = "createDay"
						r.operationGroup = ""
						r.pathPattern = "/api/days"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "DELETE":
							r.name = DeleteDayOperation
							r.summary = "日エントリを削除する"
							r.operationID = "deleteDay"
							r.operationGroup = ""
							r.pathPattern = "/api/days/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "PUT":
							r.name = UpdateDayOperation
							r.summary = "日エントリを更新する"
							r.operationID = "updateDay"
							r.operationGroup = ""
							r.pathPattern = "/api/days/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}

				}

			case 'o': // Prefix: "oidc-config"

				if l := len("oidc-config"); len(elem) >= l && elem[0:l] == "oidc-config" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetOIDCConfigOperation
						r.summary = "OIDC クライアント設定を取得する（認証不要）"
						r.operationID = "getOIDCConfig"
						r.operationGroup = ""
						r.pathPattern = "/api/oidc-config"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 's': // Prefix: "shopping"

				if l := len("shopping"); len(elem) >= l && elem[0:l] == "shopping" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "POST":
						r.name = CreateShoppingItemOperation
						r.summary = "買い物アイテムを追加する"
						r.operationID = "createShoppingItem"
						r.operationGroup = ""
						r.pathPattern = "/api/shopping"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'r': // Prefix: "reorder"
						origElem := elem
						if l := len("reorder"); len(elem) >= l && elem[0:l] == "reorder" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "POST":
								r.name = ReorderShoppingItemsOperation
								r.summary = "買い物アイテムの順序を変更する"
								r.operationID = "reorderShoppingItems"
								r.operationGroup = ""
								r.pathPattern = "/api/shopping/reorder"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

						elem = origElem
					}
					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "DELETE":
							r.name = DeleteShoppingItemOperation
							r.summary = "買い物アイテムを削除する"
							r.operationID = "deleteShoppingItem"
							r.operationGroup = ""
							r.pathPattern = "/api/shopping/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "PUT":
							r.name = UpdateShoppingItemOperation
							r.summary = "買い物アイテムを更新する"
							r.operationID = "updateShoppingItem"
							r.operationGroup = ""
							r.pathPattern = "/api/shopping/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}

				}

			case 'u': // Prefix: "userinfo"

				if l := len("userinfo"); len(elem) >= l && elem[0:l] == "userinfo" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetUserInfoOperation
						r.summary = "ログイン中のユーザー情報を取得する"
						r.operationID = "getUserInfo"
						r.operationGroup = ""
						r.pathPattern = "/api/userinfo"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}