# backend/internal/database/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlcgen

import (
	"time"
)

type Block struct {
	ID         string     `json:"id"`
	CreatedAt  time.Time  `json:"created_at"`
	ArchivedAt *time.Time `json:"archived_at"`
}

type DayEntry struct {
	ID      string  `json:"id"`
	BlockID string  `json:"block_id"`
	Date    string  `json:"date"`
	Note    *string `json:"note"`
	Lunch   *string `json:"lunch"`
	Dinner  *string `json:"dinner"`
}

type ShoppingItem struct {
	ID        string    `json:"id"`
	BlockID   string    `json:"block_id"`
	Name      string    `json:"name"`
	Category  string    `json:"category"`
	Note      *string   `json:"note"`
	SortOrder int64     `json:"sort_order"`
	Checked   int64     `json:"checked"`
	CreatedAt time.Time `json:"created_at"`
}